mongo/requirements.txt
aioredis==1.3.1
arrow==1.2.3
asgiref==3.6.0
async-timeout==4.0.2
attrs==20.1.0
autobahn==20.7.1
Automat==20.2.0
bench==2.8
binaryornot==0.4.4
certifi==2022.9.24
cffi==1.15.1
channels
channels-redis==3.0.1
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
constantly==15.1.0
contourpy==1.0.7
cookiecutter==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.1
cycler==0.11.0
daphne
dataclasses==0.6
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.1.8
django-allauth==0.52.0
django-cors-headers==3.13.0
django-csp==3.7
django-environ==0.10.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djongo==1.2.31
djongox==1.3.6.3
dnspython==2.2.1
drf-yasg==1.21.4
drf-yasg2==1.19.4
environ==1.0
et-xmlfile==1.1.0
excel2json-3==0.1.6
fonttools==4.39.3
fpdf==1.7.2
frappe-bench==5.16.4
freetype-py==2.3.0
gitdb==4.0.10
GitPython==3.1.32
gunicorn==20.0.4
hiredis==2.2.3
honcho==1.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jinja2-time==0.2.0
JPype1==1.4.1
kiwisolver==1.4.4
lml==0.1.0
lxml==4.9.2
MarkupSafe==2.1.1
matplotlib==3.7.1
msgpack==1.0.0
numpy==1.24.1
oauthlib==3.2.2
openapi-codec==1.3.2
openpyxl==3.0.10
packaging==22.0
pandas==1.5.2
pdfkit==1.0.0
Pillow==9.3.0
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.23.0
pycparser==2.20
pyexcel==0.7.0
pyexcel-ezodf==0.3.4
pyexcel-io==0.6.6
pyexcel-ods3==0.6.1
pyexcel-xlsx==0.6.0
PyExcelerate==0.10.0
PyHamcrest==2.0.2
PyJWT==1.7.1
pymongo==3.12.3
pyOpenSSL==19.1.0
pyparsing==3.0.9
pypiwin32==223
python-bsonjs==0.3.0
python-crontab==2.6.0
python-dateutil==2.8.2
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2022.6
pywin32==306
PyYAML==6.0
redis==4.5.4
reportlab==3.6.13
requests==2.28.1
requests-oauthlib==1.3.1
rest-swagger==1.0.0
rlPyCairo==0.2.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
semantic-version==2.8.5
service-identity==18.1.0
simplejson==3.18.0
six==1.15.0
smmap==5.0.0
sqlparse==0.4.2
text-unidecode==1.3
texttable==1.6.7
tomli==2.0.1
Twisted==21.2.0
twisted-iocpsupport==1.0.3
txaio==20.4.1
typing_extensions==4.5.0
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.13
wkhtmltopdf==0.2
xlrd==2.0.1
xlsx2pdf==1.0.4
zope.interface==5.1.0
